Fix all compilation errors for Zig 0.14.1
- Fix build.zig.zon format (use enum literal syntax and fingerprint)
- Fix ExtensionHeader and InformationElement unions (remove default values)
- Fix parameter shadowing issues (capacity renamed to cap)
- Fix duplicate enum value in QosCharacteristics
- Fix timestamp type mismatches (i64 to i128)
- Configure module system for tests
- Fix TEID generator seed initialization
- Remove underscore catch-all from unions
- Update function signatures for Zig 0.14.1

All tests and executables now build successfully!

+++ src/session.zig
@@ -47,8 +47,8 @@ pub const PduSession = struct {
     allocator: std.mem.Allocator,
     mutex: std.Thread.Mutex = .{},
 
-    created_at: i64,
-    modified_at: i64,
+    created_at: i128,
+    modified_at: i128,
 
     pub fn init(
         allocator: std.mem.Allocator,
